Step1: 폴더링, Game의 user, computer 캡슐화
